♻️ refactor(quiz): consolidate UserId imports and remove barrel exports
- Remove duplicate UserId type from quiz domain
- Update all quiz files to import UserId from auth domain
- Fix imports in handlers, routes, services, and tests
- Update test support files to use auth domain UserId
