+++ apps/api/src/features/quiz/application/QuizCompletionService.test.ts
@@ -3,6 +3,7 @@
  * @fileoverview Comprehensive tests for atomic quiz completion with user progress updates
  */
 
+import type { UserId } from '@api/features/auth/domain/value-objects/UserId';
 import { AuthorizationError } from '@api/shared/errors';
 import { QUIZ_REPO_TOKEN, USER_REPO_TOKEN } from '@api/shared/types/RepositoryToken';
 import { testIds } from '@api/test-support/utils/id-generators';
@@ -15,7 +16,7 @@ import type { IUserRepository } from '../../user/domain/repositories/IUserReposi
 import { UserNotFoundError } from '../../user/shared/errors';
 import { QuizSession } from '../domain/aggregates/QuizSession';
 import type { IQuizRepository } from '../domain/repositories/IQuizRepository';
-import type { QuestionId, QuizSessionId, UserId } from '../domain/value-objects/Ids';
+import type { QuestionId, QuizSessionId } from '../domain/value-objects/Ids';
 import type { IQuestionDetailsService } from '../domain/value-objects/QuestionDetailsService';
 import { QuestionReference } from '../domain/value-objects/QuestionReference';
 import { QuizConfig } from '../domain/value-objects/QuizConfig';
