+++ apps/api/src/features/quiz/get-results/handler.test.ts
@@ -1,8 +1,9 @@
+import type { UserId } from '@api/features/auth/domain/value-objects/UserId';
 import { testIds } from '@api/test-support/utils/id-generators';
 import { TestClock } from '@api/test-support/utils/TestClock';
 import { beforeEach, describe, expect, it, vi } from 'vitest';
 import { QuizSession } from '../domain/aggregates/QuizSession';
-import type { OptionId, QuestionId, QuizSessionId, UserId } from '../domain/value-objects/Ids';
+import type { OptionId, QuestionId, QuizSessionId } from '../domain/value-objects/Ids';
 import type { QuestionDetails } from '../domain/value-objects/QuestionDetailsService';
 import { QuestionReference } from '../domain/value-objects/QuestionReference';
 import { QuizConfig } from '../domain/value-objects/QuizConfig';
