+++ apps/api/src/features/quiz/get-results/route.ts
@@ -3,6 +3,7 @@
  * @fileoverview HTTP endpoint for retrieving quiz results and scoring using route utilities
  */
 
+import { UserId } from '@api/features/auth/domain/value-objects/UserId';
 import { getRepositoryFromContext } from '@api/infra/repositories/providers';
 import type { AuthUser } from '@api/middleware/auth/auth-user';
 import type { LoggerVariables } from '@api/middleware/logger';
@@ -15,7 +16,7 @@ import { QUIZ_REPO_TOKEN } from '@api/shared/types/RepositoryToken';
 import { isValidUUID } from '@api/shared/validation/constants';
 import { Hono } from 'hono';
 import type { IQuizRepository } from '../domain/repositories/IQuizRepository';
-import { QuizSessionId, UserId } from '../domain/value-objects/Ids';
+import { QuizSessionId } from '../domain/value-objects/Ids';
 import type { StubQuestionDetailsService } from '../domain/value-objects/QuestionDetailsService';
 import { QuizDependencyProvider } from '../shared/dependencies';
 import { mapGetResultsError } from '../shared/error-mapper';
